+++ PgRoutiner/Builder/CodeBuilders/UnitTests/TestFixtures.cs
@@ -18,7 +18,11 @@ public class TestFixtures : Code
         {
             foreach (var line in settings.SourceHeaderLines)
             {
-                sb.AppendLine(string.Format(line, DateTime.Now));
+                var value = string.Format(line, DateTime.Now);
+                if (!value.StartsWith("#pragma"))
+                {
+                    sb.AppendLine(value);
+                }
             }
         }
         if (globalUsings == null)
@@ -196,15 +200,19 @@ public class TestFixtures : Code
         sb.AppendLine(@$"{I2}}}");
         sb.AppendLine(@$"{I1}}}");
         sb.AppendLine(@"");
+
         sb.AppendLine(@$"{I1}[CollectionDefinition(""PostgreSqlDatabase"")]");
         sb.AppendLine(@$"{I1}public class DatabaseFixtureCollection : ICollectionFixture<PostgreSqlFixture> {{ }}");
-        sb.AppendLine(@"");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I1}/// <summary>");
+        sb.AppendLine(@$"{I1}/// PostgreSQL Unit Test Fixture using configuration settings from testsettings.json");
+        sb.AppendLine(@$"{I1}/// </summary>");
         sb.AppendLine(@$"{I1}[Collection(""PostgreSqlDatabase"")]");
-        sb.AppendLine(@$"{I1}public abstract class PostgreSqlUnitTestFixture : IDisposable");
+        sb.AppendLine(@$"{I1}public abstract class PostgreSqlConfigurationFixture : IDisposable");
         sb.AppendLine(@$"{I1}{{");
         sb.AppendLine(@$"{I2}protected NpgsqlConnection Connection {{ get; }}");
-        sb.AppendLine(@"");
-        sb.AppendLine(@$"{I2}protected PostgreSqlUnitTestFixture(PostgreSqlFixture fixture)");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I2}protected PostgreSqlConfigurationFixture(PostgreSqlFixture fixture)");
         sb.AppendLine(@$"{I2}{{");
         sb.AppendLine(@$"{I3}if (Config.Value.UnitTestsNewDatabaseFromTemplate)");
         sb.AppendLine(@$"{I3}{{");
@@ -220,13 +228,13 @@ public class TestFixtures : Code
         sb.AppendLine(@$"{I4}Connection = fixture.Connection.CloneWith(fixture.Connection.ConnectionString);");
         sb.AppendLine(@$"{I4}Connection.Open();");
         sb.AppendLine(@$"{I3}}}");
-        sb.AppendLine(@"");
+        sb.AppendLine("");
         sb.AppendLine(@$"{I3}if (Config.Value.UnitTestsUnderTransaction)");
         sb.AppendLine(@$"{I3}{{");
         sb.AppendLine(@$"{I4}Connection.Execute(""begin"");");
         sb.AppendLine(@$"{I3}}}");
         sb.AppendLine(@$"{I2}}}");
-        sb.AppendLine(@"");
+        sb.AppendLine("");
         sb.AppendLine(@$"{I2}[System.Diagnostics.CodeAnalysis.SuppressMessage(""Usage"", ""CA1816:Dispose methods should call SuppressFinalize"", Justification = ""XUnit"")]");
         sb.AppendLine(@$"{I2}public void Dispose()");
         sb.AppendLine(@$"{I2}{{");
@@ -243,6 +251,76 @@ public class TestFixtures : Code
         sb.AppendLine(@$"{I3}}}");
         sb.AppendLine(@$"{I2}}}");
         sb.AppendLine(@$"{I1}}}");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I1}/// <summary>");
+        sb.AppendLine(@$"{I1}/// PostgreSQL Unit Test Fixture that uses a pre-created test database.");
+        sb.AppendLine(@$"{I1}/// </summary>");
+        sb.AppendLine(@$"{I1}[Collection(""PostgreSqlDatabase"")]");
+        sb.AppendLine(@$"{I1}public abstract class PostgreSqlTestDatabaseFixture : IDisposable");
+        sb.AppendLine(@$"{I1}{{");
+        sb.AppendLine(@$"{I2}protected NpgsqlConnection Connection {{ get; }}");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I2}protected PostgreSqlTestDatabaseFixture(PostgreSqlFixture fixture)");
+        sb.AppendLine(@$"{I2}{{");
+        sb.AppendLine(@$"{I3}Connection = fixture.Connection.CloneWith(fixture.Connection.ConnectionString);");
+        sb.AppendLine(@$"{I3}Connection.Open();");
+        sb.AppendLine(@$"{I2}}}");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I2}[System.Diagnostics.CodeAnalysis.SuppressMessage(""Usage"", ""CA1816:Dispose methods should call SuppressFinalize"", Justification = ""XUnit"")]");
+        sb.AppendLine(@$"{I2}public virtual void Dispose()");
+        sb.AppendLine(@$"{I2}{{");
+        sb.AppendLine(@$"{I3}Connection.Close();");
+        sb.AppendLine(@$"{I3}Connection.Dispose();");
+        sb.AppendLine(@$"{I2}}}");
+        sb.AppendLine(@$"{I1}}}");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I1}/// <summary>");
+        sb.AppendLine(@$"{I1}/// PostgreSQL Unit Test Fixture uses a pre-created test database that runs each test under a new transaction that is rolled-back automatically.");
+        sb.AppendLine(@$"{I1}/// </summary>");
+        sb.AppendLine(@$"{I1}[Collection(""PostgreSqlDatabase"")]");
+        sb.AppendLine(@$"{I1}public abstract class PostgreSqlTestDatabaseTransactionFixture : PostgreSqlTestDatabaseFixture, IDisposable");
+        sb.AppendLine(@$"{I1}{{");
+        sb.AppendLine(@$"{I2}protected PostgreSqlTestDatabaseTransactionFixture(PostgreSqlFixture fixture) : base(fixture)");
+        sb.AppendLine(@$"{I2}{{");
+        sb.AppendLine(@$"{I3}Connection.Execute(""begin"");");
+        sb.AppendLine(@$"{I2}}}");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I2}[System.Diagnostics.CodeAnalysis.SuppressMessage(""Usage"", ""CA1816:Dispose methods should call SuppressFinalize"", Justification = ""XUnit"")]");
+        sb.AppendLine(@$"{I2}public override void Dispose()");
+        sb.AppendLine(@$"{I2}{{");
+        sb.AppendLine(@$"{I3}Connection.Execute(""rollback"");");
+        sb.AppendLine(@$"{I3}base.Dispose();");
+        sb.AppendLine(@$"{I2}}}");
+        sb.AppendLine(@$"{I1}}}");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I1}/// <summary>");
+        sb.AppendLine(@$"{I1}/// PostgreSQL Unit Test Fixture using a a database that is created from the test database as a template for the each new tests and cleaned-up (dropped) after the test.");
+        sb.AppendLine(@$"{I1}/// </summary>");
+        sb.AppendLine(@$"{I1}[Collection(""PostgreSqlDatabase"")]");
+        sb.AppendLine(@$"{I1}public abstract class PostgreSqlTestTemplateDatabaseFixture : IDisposable");
+        sb.AppendLine(@$"{I1}{{");
+        sb.AppendLine(@$"{I2}protected NpgsqlConnection Connection {{ get; }}");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I2}protected PostgreSqlTestTemplateDatabaseFixture(PostgreSqlFixture fixture)");
+        sb.AppendLine(@$"{I2}{{");
+        sb.AppendLine(@$"{I3}var dbName = string.Concat(Config.Value.TestDatabaseName, ""_"", Guid.NewGuid().ToString().Replace(""-"", ""_""));");
+        sb.AppendLine(@$"{I3}using var connection = new NpgsqlConnection(Config.ConnectionString);");
+        sb.AppendLine(@$"{I3}PostgreSqlFixture.CreateDatabase(connection, dbName, connection.Database);");
+        sb.AppendLine(@$"{I3}Connection = fixture.Connection.CloneWith(fixture.Connection.ConnectionString);");
+        sb.AppendLine(@$"{I3}Connection.Open();");
+        sb.AppendLine(@$"{I3}Connection.ChangeDatabase(dbName);");
+        sb.AppendLine(@$"{I2}}}");
+        sb.AppendLine("");
+        sb.AppendLine(@$"{I2}[System.Diagnostics.CodeAnalysis.SuppressMessage(""Usage"", ""CA1816:Dispose methods should call SuppressFinalize"", Justification = ""XUnit"")]");
+        sb.AppendLine(@$"{I2}public virtual void Dispose()");
+        sb.AppendLine(@$"{I2}{{");
+        sb.AppendLine(@$"{I3}Connection.Close();");
+        sb.AppendLine(@$"{I3}Connection.Dispose();");
+        sb.AppendLine(@$"{I3}using var connection = new NpgsqlConnection(Config.ConnectionString);");
+        sb.AppendLine(@$"{I3}PostgreSqlFixture.DropDatabase(connection, Connection.Database);");
+        sb.AppendLine(@$"{I2}}}");
+        sb.AppendLine(@$"{I1}}}");
+
         if (!settings.UseFileScopedNamespaces)
         {
             sb.AppendLine(@"}");